feat: add decomposition for pagination
